remove evidence form display for now
Fixes #26.

It turns out that there was some funniness going on in the widgets for
generating a table. genTable(), in widgets.js builds tables from
scratch in bare html, not escaping anything. These tables are then
passed into mustache essentially untouched, as indicated by the triple
curly bracket notation - ie.,

{{{simModelTable}}}

this is a problem for the URL labels that contain characters reserved
in HTML, such as angle brackets.

One approach might be to back away from the triple curlies, and to let
mustache esacpe things. This will not work, as we are sending legal
html of forms like "<A HREF=..." into the template. We can't have this
escaped.

To get around this, I added convChars() to widgets.js, to convert all
escapd characters to an html friendly form - "<" becomes "&lt;",
etc. This is now called for all HREF labels and free text generated in
widgets.js.

One minor complication of this approach. Some calls generate HTML that
has already been escaped. Naive escaping will muck things up- changing
"&lt;" to "&amp lt;", for example. convChars() handles this by
changing the ampersand only if it is not prefixing an already-escaped
character. In other words, "&" will become "&amp;", but "&lt;" will
not be changed.