Fixed hookup of grid view to analysis results.
Standardized on target_species at http param.

Analyze.js needs to send the species parameter to modelinfo;
it now gets this from a div element that is separated from the
main form to avoid confusion (note the value of this will always
be for the **current** page; we want to avoid sending this
as a form param as it would prevent the user from selecting new
values).

This almost fixes issue #134