#426 fetches genes that are associated to diseases and builds gene_associations object
Now calls fetchDiseaseGeneAssociations on every disease associated with the phenotype.

My main problem with this is the load time is ridiculously long. In the case of /phenotype/MP_0001322, we had to fetch the genes associated with 412 diseases, and now the gene_associations table is 1838 elements long.

I am considering cutting off the results after 1000. I may also need to use a different fetch function.