Updates clean.sh to remove dist/.
Updates gulpfile.js to use browserify. The resulting bundle is not currently being used. This is a work-in-progress.

Disable update_dependencies.sh, since it is no longer necessary.

Fixes install.sh to either npm link or npm install, depending upon the PHENOGRID_LOCAL_DEV environment symbol being defined.

Use non-jQuery interface to Phenogrid. Instead of:
	$("#phen_vis").phenogrid({phenotypeData: phenotype_list});
we use:
	var phenogridContainer = document.getElementById('phen_vis');
	var phenogridOpts = {
	                        phenotypeData: phenotype_list
	                    };

	Phenogrid.createPhenogridForElement(phenogridContainer, phenogridOpts);

Many changes to webapp.js so that phenogrid resources are pulled from node_modules/phenogrid instead of widgets/phenogrid.

Updates package.json license: field to be 'CC-BY-3.0' instead of 'BSD'. This eliminates an NPM warning and corresponds to the notes in the footer of the Monarch app.

Update phenogrid.mustache to use node_modules/phenogrid instead of widgets/

Deletes the /widgets/phenogrid directory.

Developer Notes
===============

I’ve modified install.sh so that you can define an environment variable to enable the use of the local copy. If a subsequent `npm install` detects that a `npm link`-ed package is in place, then it will leave it alone. So if our phenogrid repo is a sibling of monarchapp, we can define:

	export PHENOGRID_LOCAL_DEV=../phenogrid

Inside of install.sh, there’s now code that looks like:
		...
if [ $PHENOGRID_LOCAL_DEV ]; then
	echo "Using local phenogrid from: $PHENOGRID_LOCAL_DEV"
	npm link $PHENOGRID_LOCAL_DEV
fi
npm install
	...

This way we get a local override if we want it, but otherwise pull the standard version (assuming we’ve published that to npm).

Highlights of the changes:

- The phenogrid build process is simplified; there is no need for a separate bundle prefixed by jQuery. I also got rid of the unnecessary browserify-shim code.

- The index.html that serves as an example of phenogrid and that is also used in monarch-app (via iframe) now uses standard HTML and does not require jQuery.

- It turns out that most of the problems and kludges were required to export jQuery from phenogrid. But there is no need to do so. Not exporting jQuery will also address the issue of supporting modern frameworks like React and Angular, where jQuery is unnecessary and therefore undesirable.

- We are currently NOT bundling (via browserify) the monarch-app resources via browserify yet, but we are closer than before. The gulpfile and package.json are doing the bundling, but the server is not serving the bundled monarch. This is an optimization we can pursue once the phenogrid code duplication is eliminated.

- monarch-app now ‘npm install’s phenogrid and uses the resulting tree of files from node_modules/phenogrid, rather than the /widgets/phenogrid code that is in GitHub (soon to be deprecated).

- As an alternative to `npm install`, local phenogrid developers can use the `npm link` command to point from their monarch-app to their local phenogrid. That way changes to phenogrid can be seen immediately (maybe a server restart is required; haven’t verified this yet). NPM docs related to publish and link are:
	https://docs.npmjs.com/misc/developers
	https://docs.npmjs.com/cli/link

- Sinopia is NOT required. It IS useful for testing `npm install`, though.

- /widgets/phenogrid is REMOVED from monarch-app.