One goal behind this PR is to use NPM-installed most-recent versions of software modules that are currently stored within the git repo and are wildly out of date. This will ease maintenance and understanding of the codebase, as well as enabling more rapid development.
- Without loss of functionality (based upon my eyeballs and tests), I was
able to eliminate the monarch-app version of pup-tent in favor of the
NPM-available version.
- Similarly, I replaced the hardcoded d3.js with an NPM version.
- I deleted raphael.js, which was apparently not used.

I got ambitious and eliminated the old stale 3.0.0 bootstrap.min.js in
favor of the current 3.3.5 version. Unfortunately, this produced some
undesirable UI effects that I have been working through. The usual
cause is that we had CSS workarounds for 3.0.0 behaviors and these
workarounds are either no longer necessary or are inappropriate for
3.3.5.

Known problems related to the Bootstrap upgrade are:
- Navbar search form is a few pixels off
- The 'tabs' within the particular gene/disease/phenotype page are a few
  pixels off of their baseline, and not fonted correctly.
- Alignment issues within the Analyze Phenotypes page

I plan on fixing these in a subsequent PR.

---

Other changes:

When in non-production mode, an error page produces a visible stack trace and
error message, instead of hiding it in the HTML comment.